custom error handling - brand new custom error handler - dedicated "wormhole collapsed" error page - in production set LM_DEBUG to FALSE; error page will only show unique error ID - lookup the ID in syslog to know the details - in development, set LM_DEBUG to TRUE; error page will show the ID and stack trace - non-admin users will no longer see poller and ESI error notifications

+++ include/checkpoller.php
@@ -1,7 +1,7 @@
 <?php
 	checksession(); //check if we are called by a valid session
 	$LOCKFILE="../var/poller.lock";
-	if (file_exists($LOCKFILE) && ((time()-filemtime($LOCKFILE)) > 900)) {
+	if (file_exists($LOCKFILE) && ((time()-filemtime($LOCKFILE)) > 900) && checkrights("Administrator")) {
 		echo('<div class="newmsg"><img src="'.getUrl().'img/exc.gif" alt="Warning" /> EVE API Poller has hanged, contact Administrator!');
 		if (checkrights("Administrator")) {
 			echo('<button type="button" onclick="location.href=\'index.php?id=5&id2=9\'">Reset poller</button>');
